% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{mean_binomial}
\alias{mean_binomial}
\title{mean_binomial}
\usage{
mean_binomial(n, p)
}
\arguments{
\item{n}{-> number of trials}

\item{p}{-> the probability of success}
}
\value{
the mean or expected value
}
\description{
function that calculates the mean of outcomes, or the
expected value.
}
