% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{power}
\alias{power}
\title{simple function for cleaner roots.}
\usage{
power(value, power)
}
\arguments{
\item{value}{the value to apply root to.}

\item{power}{the value to raise value to.}
}
\value{
the new value raised to the power.
}
\description{
simple function for cleaner roots.
}
\examples{
power(2, 2)
> 4

power(2, 3)
> 8

}
