% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{pr.binomial}
\alias{pr.binomial}
\title{pr.binomial}
\usage{
pr.binomial(n, j, p)
}
\arguments{
\item{n}{\itemize{
\item the number of trials
}}

\item{j}{\itemize{
\item the number of events to occur
}}

\item{p}{\itemize{
\item the probability of j
}}
}
\value{
probability of Binomial Random Variable
}
\description{
Function for calculating probability for Bernouli Trials.
}
\examples{
pr.binomial(10, 2, 1/6)
> 0.29071

pr.binomial(n = 5, j = 5, p = 1/2)
> 0.03125

}
