% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basics.R
\name{some_function}
\alias{some_function}
\title{This is a test function.}
\usage{
some_function(x)
}
\arguments{
\item{x}{-> integer}
}
\value{
Nothing
}
\description{
It just loops through and prints values based on integer input.
}
