% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{var.binomial}
\alias{var.binomial}
\title{var.binomial}
\usage{
var.binomial(n, p)
}
\arguments{
\item{n}{-> number of trials}

\item{p}{-> the probability of success}
}
\value{
the mean or expected value
}
\description{
function that calculates the variance of outcomes.
}
