% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomial.R
\name{var_binomial}
\alias{var_binomial}
\title{var_binomial}
\usage{
var_binomial(n, p)
}
\arguments{
\item{n}{-> number of trials}

\item{p}{-> the probability of success}
}
\value{
the variance of trials
}
\description{
function that calculates the variance of outcomes.
}
